Move water spawning code.
