+++ rogue/src/map_builders/simple_map.rs
@@ -29,7 +29,6 @@ impl MapBuilder for SimpleMapBuilder {
         self.map.clone()
     }
 
-
     fn build_map(&mut self) -> terrain_spawners::water::WaterSpawnTable {
         self.rooms_and_corridors();
 
@@ -73,29 +72,8 @@ impl MapBuilder for SimpleMapBuilder {
         }
     }
 
-    fn spawn_water(&mut self, ecs: &mut World, water_spawn_table: &WaterSpawnTable) {
-        for e in &water_spawn_table.shallow {
-            {
-                let mut map = ecs.write_resource::<Map>();
-                let idx = map.xy_idx(e.x, e.y);
-                if map.blocked[idx] {
-                    continue;
-                }
-                map.ok_to_spawn[idx] = true;
-            }
-            terrain_spawners::water::shallow_water(ecs, e.x, e.y, e.fgcolor, e.bgcolor);
-        }
-        for e in &water_spawn_table.deep {
-            {
-                let mut map = ecs.write_resource::<Map>();
-                let idx = map.xy_idx(e.x, e.y);
-                if map.blocked[idx] {
-                    continue;
-                }
-                map.ok_to_spawn[idx] = true;
-            }
-            terrain_spawners::water::deep_water(ecs, e.x, e.y, e.fgcolor, e.bgcolor);
-        }
+    fn spawn_water(&mut self, ecs: &mut World, water_spawn_table: &terrain_spawners::water::WaterSpawnTable) {
+        terrain_spawners::water::spawn_water_from_table(ecs, water_spawn_table)
     }
 
     fn spawn_terrain(&mut self, ecs: &mut World) {
